requirements.txt
numpy>=1.21.0
pandas>=1.3.0
requests>=2.26.0
python-dotenv>=0.19.0 